requirements.txt
langchain
langchain-openai
langchain-community
openai
faiss-cpu
python-dotenv
fastapi
uvicorn
pydantic
requests
tiktoken
python-jose[cryptography]
azure-ai-inference
azure-cosmos
spacy
azure-storage-blob
